agents/python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rukkie-agent"
version = "0.1.0"
description = "RukkiePulse agent SDK for Python"
requires-python = ">=3.9"
dependencies = [
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-flask>=0.45b0",
    "opentelemetry-instrumentation-requests>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.100", "starlette>=0.27"]
flask = ["flask>=2.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["rukkie_agent*"]
